Added minimal UI Main Menu
Added an intro level with a main menu. When you press on Start game it opens up the main sandbox level. Have to add in the functionality for the other buttons. Basically just other widgets that would have credits and settings functions. Figured I'd wait until there is an ability for 2D artwork for that.